FAIR-2507 拆分文件夹
