+++ src/app/net/pull.ts
@@ -2,20 +2,22 @@
  * 主动向后端通讯
  */
 import { uploadFileUrl } from '../config';
-import { callRequestAsync, callRequestAsyncNeedLogin,getStoreData, setStoreData } from '../middleLayer/wrap';
-import { CloudCurrencyType } from '../publicLib/interface';
-import { getModulConfig } from '../publicLib/modulConfig';
-import { piFetch, unicodeArray2Str } from '../publicLib/tools';
-import { largeUnit2SmallUnit } from '../publicLib/unitTools';
+import { getModulConfig, PAGELIMIT } from '../public/config';
+import { CloudCurrencyType } from '../public/interface';
+import { getStore, setStore } from '../store/memstore';
+// tslint:disable-next-line:max-line-length
+import { parseCloudAccountDetail, parseCloudBalance, parseConvertLog, parseDividHistory, parseExchangeDetail, parseMiningRank, parseMyInviteRedEnv, parseSendRedEnvLog, splitCloudCurrencyDetail } from '../utils/parse';
+import { base64ToFile, getUserInfo, piFetch, popNewMessage, unicodeArray2Str } from '../utils/pureUtils';
 import { showError } from '../utils/toolMessages';
-import { base64ToFile, getUserInfo, popNewMessage } from '../utils/tools';
+import { largeUnit2SmallUnit } from '../utils/unitTools';
+import { requestAsync } from './login';
 
 /**
  * 获取指定类型的货币余额
  */
 export const getBalance = async (currencyType: CloudCurrencyType) => {
     const msg = { type: 'wallet/account@get', param: { list: `[${currencyType}]` } };
-    callRequestAsync(msg).then(r => {
+    requestAsync(msg).then(r => {
         // todo 这里更新余额
     });
 };
@@ -28,7 +30,7 @@ export const getBalance = async (currencyType: CloudCurrencyType) => {
 export const inputInviteCdKey = async (code) => {
     const msg = { type: 'wallet/cloud@input_cd_key', param: { code: code } };
     try {
-        const res = await callRequestAsync(msg);
+        const res = await requestAsync(msg);
 
         return [res.fuid];
     } catch (err) {
@@ -61,7 +63,7 @@ export const  sendRedEnvlope = async (rtype: string, ctype: number, totalAmount:
     };
 
     try {
-        const res = await callRequestAsyncNeedLogin(msg,secretHash);
+        const res = await requestAsync(msg);
 
         return res.value;
     } catch (err) {
@@ -79,7 +81,7 @@ export const takeRedBag = async (rid) => {
     
     try {
         // tslint:disable-next-line:no-unnecessary-local-variable
-        const res = await callRequestAsync(msg);
+        const res = await requestAsync(msg);
 
         return res;
     } catch (err) {
@@ -97,7 +99,7 @@ export const convertRedBag = async (cid) => {
 
     try {
         // tslint:disable-next-line:no-unnecessary-local-variable
-        const res = await callRequestAsync(msg);
+        const res = await requestAsync(msg);
 
         return res;
     } catch (err) {
@@ -119,7 +121,7 @@ export const queryRedBagDesc = async (cid: string) => {
         }
     };
 
-    return callRequestAsync(msg);
+    return requestAsync(msg);
 };
 
 // ==========================================红包end
@@ -132,7 +134,7 @@ export const getAward = async () => {
 
     try {
         // tslint:disable-next-line:no-unnecessary-local-variable
-        const detail = await callRequestAsync(msg);
+        const detail = await requestAsync(msg);
         
         return detail;
         
@@ -149,7 +151,7 @@ export const getAward = async () => {
 export const setData = async (param) => {
     const msg = { type: 'wallet/data@set', param: param };
 
-    return callRequestAsync(msg);
+    return requestAsync(msg);
 };
 
 /**
@@ -158,7 +160,7 @@ export const setData = async (param) => {
 export const getData = async (key) => {
     const msg = { type: 'wallet/data@get', param: { key } };
 
-    return callRequestAsync(msg);
+    return requestAsync(msg);
 };
 
 /**
@@ -173,7 +175,7 @@ export const getUserList = async (uids: number[], isOpenid?: number) => {
     }
 
     try {
-        const res = await callRequestAsync(msg);
+        const res = await requestAsync(msg);
         if (res.value[0]) {
             const resAry = [];
             for (const element of res.value) {
@@ -195,7 +197,7 @@ export const getUserList = async (uids: number[], isOpenid?: number) => {
 export const doChat = async () => {
     const msg = { type: 'wallet/cloud@chat', param: {} };
 
-    callRequestAsync(msg).then(r => {
+    requestAsync(msg).then(r => {
         // 通信成功
     });
 };
@@ -206,7 +208,7 @@ export const doChat = async () => {
 export const verifyPhone = async (phone:string,num: string) => {
     const msg = { type: 'wallet/user@check_phone', param: { phone,num } };
     try {
-        await callRequestAsync(msg); 
+        await requestAsync(msg); 
 
         return false;
     } catch (err) {
@@ -230,7 +232,7 @@ export const sendCode = async (phone: string, num: string,verify:boolean = true)
     }
     const msg = { type: 'wallet/sms@send_sms_code', param: { phone, num, name: getModulConfig('WALLET_NAME') } };
     try {
-        return await callRequestAsync(msg);
+        return await requestAsync(msg);
     } catch (err) {
         showError(err && (err.result || err.type));
 
@@ -252,7 +254,7 @@ export const regPhone = async (phone: string, num:string, code: string) => {
     const msg = { type: 'wallet/user@reg_phone', param: { phone, old_phone, code,num,old_num } };
     
     try {
-        return await callRequestAsync(msg);
+        return await requestAsync(msg);
     } catch (err) {
         showError(err && (err.result || err.type));
 
@@ -270,7 +272,7 @@ export const checkPhoneCode = async (phone: string, code: string,cmd?:string) =>
     }
     const msg = { type: 'wallet/user@check_phoneCode', param };
     try {
-        return await callRequestAsync(msg);
+        return await requestAsync(msg);
     } catch (err) {
         showError(err && (err.result || err.type));
 
@@ -285,7 +287,7 @@ export const unbindPhone = async (phone: string, code: string,num:string) => {
     const param:any = { phone, code,num };
     const msg = { type: 'wallet/user@unset_phone', param };
     try {
-        return await callRequestAsync(msg);
+        return await requestAsync(msg);
     } catch (err) {
         showError(err && (err.result || err.type));
 
@@ -299,33 +301,10 @@ export const unbindPhone = async (phone: string, code: string,num:string) => {
 export const getProxy = async () => {
     const msg = { type: 'wallet/proxy@get_proxy', param: {} };
 
-    return callRequestAsync(msg);
+    return requestAsync(msg);
 };
 
 // ===============================充值提现
-
-/**
- * 赎回理财产品
- */
-export const buyBack = async (timeStamp:any,secretHash:string) => {
-    const msg = {
-        type: 'wallet/manage_money@sell',
-        param: {
-            time:timeStamp
-        }
-    };
-    
-    try {
-        const res = await callRequestAsyncNeedLogin(msg,secretHash);
-        console.log('buyBack',res);
-
-        return true;
-    } catch (err) {
-        showError(err && (err.result || err.type));
-
-        return false;
-    }
-};
 
 // 上传文件
 export const uploadFile = async (base64) => {
@@ -348,9 +327,9 @@ export const uploadFile = async (base64) => {
         popNewMessage('图片上传成功');
         if (res.result === 1) {
             const sid = res.sid;
-            getStoreData('user/info').then(userInfo => {
+            getStore('user/info').then(userInfo => {
                 userInfo.avatar = sid;
-                setStoreData('user/info',userInfo);
+                setStore('user/info',userInfo);
             });
         }
     }).catch(err => {
@@ -370,7 +349,7 @@ export const changellySign = (data:any) => {
         }
     };
 
-    return callRequestAsync(msg);
+    return requestAsync(msg);
 };
 
 /**
@@ -382,7 +361,7 @@ export const getInviteUserAccIds = () => {
         param: {}
     };
 
-    return callRequestAsync(msg);
+    return requestAsync(msg);
 };
 
 /**
@@ -394,7 +373,7 @@ export const getAppleGoods = () => {
         param: {}
     };
 
-    return callRequestAsync(msg);
+    return requestAsync(msg);
 };
 
 /**
@@ -409,7 +388,7 @@ export const getUserRecentGame = (accid:number,count:number) => {
         }
     };
 
-    return callRequestAsync(msg).then(r => {
+    return requestAsync(msg).then(r => {
         if (r.list) {
             const list = [];
             r.list.forEach(v => {  // appid
@@ -419,4 +398,372 @@ export const getUserRecentGame = (accid:number,count:number) => {
             return list;
         }
     });
+};
+
+// 获取真实用户
+export const getRealUser = () => {
+    const msg = {
+        type: 'wallet/user@get_real_user',
+        param: {}
+    };
+
+    return requestAsync(msg).then(res => {
+        const userInfo  = getStore('user/info');
+        const isRealUser = res.value !== 'false';
+        
+        if (isRealUser !== userInfo.isRealUser) {
+            userInfo.isRealUser =  isRealUser;
+            setStore('user/info',userInfo);
+        }
+    });
+};
+
+/**
+ * 获取绑定的手机号
+ */
+export const getBindPhone = () => {
+    const msg = { type: 'wallet/user@get_phone',param: {} };
+
+    return requestAsync(msg).then(res => {
+        const userInfo  = getStore('user/info');
+        if (userInfo.phoneNumber !== res.phone || userInfo.areaCode !== res.num) {
+            userInfo.phoneNumber =  res.phone;
+            userInfo.areaCode = res.num;
+            setStore('user/info',userInfo);
+        }
+    });
+};
+
+/**
+ * 获取所有的货币余额
+ */
+export const getServerCloudBalance = () => {
+    const list = [];
+    list.push(CloudCurrencyType.KT);
+    list.push(CloudCurrencyType.SC);
+    const msg = { type: 'wallet/account@get', param: { list:`[${list}]` } };
+    
+    return requestAsync(msg).then(balanceInfo => {
+        console.log('balanceInfo', balanceInfo);
+        const cloudBalances = parseCloudBalance(balanceInfo);
+        const cloudWallets = getStore('cloud/cloudWallets');
+        for (const [key,value] of cloudBalances) {
+            const cloudWallet = cloudWallets.get(key);
+            cloudWallet.balance = value;
+        }
+        setStore('cloud/cloudWallets',cloudWallets);
+    });
+};
+
+/**
+ * 获取当前用户信息
+ */
+export const getUserInfoFromServer = (uids: [number]) => {
+    const msg = { type: 'wallet/user@get_infos', param: { list: `[${uids.toString()}]` } };
+
+    return requestAsync(msg).then(res => {
+        const userInfoStr = unicodeArray2Str(res.value[0]);
+        const localUserInfo = getStore('user/info');
+        console.log('localUserInfo ==== ',localUserInfo);
+        console.log('serverUserInfoStr ==== ',userInfoStr);
+        
+        if (userInfoStr !== JSON.stringify(localUserInfo)) {
+            const serverUserInfo = userInfoStr ? JSON.parse(userInfoStr) : {} ; 
+            console.log('serverUserInfo ==== ',serverUserInfo);
+            const userInfo = {};
+            for (const key in localUserInfo) {
+                if (!serverUserInfo[key]) {
+                    userInfo[key] = localUserInfo[key];
+                } else {
+                    userInfo[key] = serverUserInfo[key];
+                }
+            }
+            console.log('userInfo ==== ',userInfo);
+            setStore('user/info',userInfo);
+        } 
+        
+    });
+        
+};
+
+/**
+ * 设置用户基础信息
+ */
+export const setUserInfo = () => {
+    const userInfo = getStore('user/info');
+    const msg = { type: 'wallet/user@set_info', param: { value:JSON.stringify(userInfo) } };
+    
+    return requestAsync(msg);
+};
+
+/**
+ * 获取全部用户嗨豆排名列表
+ */
+export const getHighTop =  (num: number) => {
+    const msg = { type: 'wallet/cloud@get_high_top', param: { num: num } };
+
+    return requestAsync(msg).then(data => {
+        console.log('获取全部排名========================',data);
+        
+        return parseMiningRank(data);
+    });
+    
+};
+
+/**
+ * 获取指定货币流水
+ * filter（0表示不过滤，1表示过滤）
+ */
+export const getAccountDetail = (coin: string,filter:number,start = '') => {
+    const param:any = {
+        coin:CloudCurrencyType[coin],
+        start,
+        filter,
+        count:PAGELIMIT
+    };
+    if (start) {
+        param.start = start;
+    } 
+    const msg = {
+        type: 'wallet/account@get_detail',
+        param
+    };
+
+    return requestAsync(msg).then(res => {
+        const nextStart = res.start;
+        const detail = parseCloudAccountDetail(coin,res.value);
+        const splitDetail = splitCloudCurrencyDetail(detail); 
+        const canLoadMore = detail.length >= PAGELIMIT;
+        if (detail.length > 0) {
+            const cloudWallets = getStore('cloud/cloudWallets');
+            const cloudWallet = cloudWallets.get(CloudCurrencyType[coin]);
+            if (start) {
+                cloudWallet.otherLogs.list.push(...detail);
+                if (coin === CloudCurrencyType[CloudCurrencyType.SC] || coin === CloudCurrencyType[CloudCurrencyType.KT]) {
+                    cloudWallet.rechargeLogs.list.push(...splitDetail.rechangeList);
+                    cloudWallet.withdrawLogs.list.push(...splitDetail.withdrawList); 
+                }
+            } else {
+                cloudWallet.otherLogs.list = detail;
+                if (coin === CloudCurrencyType[CloudCurrencyType.SC] || coin === CloudCurrencyType[CloudCurrencyType.KT]) {
+                    cloudWallet.rechargeLogs.list = splitDetail.rechangeList;
+                    cloudWallet.withdrawLogs.list = splitDetail.withdrawList;
+                }
+            }
+                    
+            cloudWallet.otherLogs.start = nextStart;
+            cloudWallet.otherLogs.canLoadMore = canLoadMore;
+            setStore('cloud/cloudWallets',cloudWallets);
+
+            return cloudWallets;
+        }
+    });
+        
+};
+
+/**
+ * 获取ST价格
+ */
+export const getSilverPrice = (ispay:number = 0) => {
+    const msg = { type:'get_silverprice',param:{ ispay } };
+
+    return requestAsync(msg).then(resData => {
+        if (resData.result === 1) {
+            setStore('third/silver',{ price:resData.price,change:resData.change });
+        }
+
+        return resData;
+    });
+    
+};
+
+/**
+ * 查询红包兑换记录
+ */
+export const queryConvertLog = (start?:string) => {
+    let msg;
+    if (start) {
+        msg = {
+            type: 'query_convert_log',
+            param: {
+                start,
+                count: PAGELIMIT
+            }
+        };
+    } else {
+        msg = {
+            type: 'query_convert_log',
+            param: {
+                count: PAGELIMIT
+            }
+        };
+    }
+
+    return requestAsync(msg).then(detail => {
+        const data = parseConvertLog(detail,start);
+        setStore('activity/luckyMoney/exchange',data);
+
+        return data;
+    });
+};
+
+/**
+ * 获取分红历史记录
+ */
+export const getDividHistory = (start = '') => {
+    const msg = { 
+        type: 'wallet/cloud@get_bonus_info', 
+        param: {
+            start,
+            count:PAGELIMIT
+        } 
+    };
+    
+    return requestAsync(msg).then(data => {
+        const dividHistory = parseDividHistory(data);
+        setStore('activity/dividend/history', dividHistory);
+
+        return dividHistory;
+    });
+};
+
+/**
+ * 查询某个红包兑换详情
+ */
+export const queryDetailLog = (uid:number,rid: string,accId?:string) => {
+    const msg = {
+        type: 'query_detail_log',
+        param: {}
+    };
+    if (accId) {  // 与聊天通用的账户id
+        msg.param = {
+            acc_id: accId,
+            rid
+        };
+    } else {
+        msg.param = {
+            uid,
+            rid
+        };
+    }
+    if (rid === '-1') return;
+
+    return requestAsync(msg).then(detail => {
+        return parseExchangeDetail(detail.value);
+    });
+        
+};
+
+// 获取好友嗨豆排名
+export const getFriendsKTTops =  (arr:any) => {
+    const msg = { type:'wallet/cloud@finds_high_top',param:{ finds:JSON.stringify(arr) } };
+
+    return  requestAsync(msg).then(data => {
+        console.log('获取好友排名========================',data);
+
+        return parseMiningRank(data);
+    });
+};
+
+/**
+ * 获取邀请码领取明细
+ */
+export const getInviteCodeDetail = () => {
+    const msg = { type: 'wallet/cloud@get_invite_code_detail', param: {} };
+    
+    return requestAsync(msg).then(data => {
+        return parseMyInviteRedEnv(data.value);
+    });
+};
+
+/**
+ * 查询发送红包记录
+ */
+export const querySendRedEnvelopeRecord = (start?: string) => {
+    let msg;
+    if (start) {
+        msg = {
+            type: 'query_emit_log',
+            param: {
+                start,
+                count: PAGELIMIT
+            }
+        };
+    } else {
+        msg = {
+            type: 'query_emit_log',
+            param: {
+                count: PAGELIMIT
+            }
+        };
+    }
+
+    return requestAsync(msg).then(detail => {
+        const data = parseSendRedEnvLog(detail.value,start);
+        setStore('activity/luckyMoney/sends',data);
+
+        return data;
+    });
+
+};
+
+/**
+ * 获取后台发起分红历史记录
+ */
+export const getBonusHistory = () => {
+    const msg = { type:'wallet/cloud@get_bonus_history',param:{} };
+
+    return requestAsync(msg);
+};
+
+/**
+ * 获取分红汇总信息
+ */
+export const getDividend = async () => {
+    const msg = { type: 'wallet/cloud@get_bonus_total', param: {} };
+    const data = await getBonusHistory();
+    // TODO 这里进行单位转换
+    const num = (data.value !== '') ? data.value[0][1] : 0;
+    const yearIncome = (num * 365 / 7).toFixed(4); 
+    
+    return requestAsync(msg).then(data => {
+        const dividend: any = {
+            totalDivid: data.value[0],
+            totalDays: data.value[1],
+            thisDivid: data.value[2],
+            yearIncome: yearIncome
+        };
+        setStore('dividTotal', dividend);
+
+        return dividend;
+    });
+};
+
+/**
+ * 获取单个用户信息
+ */
+export const getOneUserInfo = (uids: number[], isOpenid?: number) => {
+    let msg = {};
+    if (isOpenid) {
+        msg = { type: 'wallet/user@get_infos', param: { list: `[${uids.toString()}]`, isOpenid } };
+    } else {
+        msg = { type: 'wallet/user@get_infos', param: { list: `[${uids.toString()}]` } };
+    }
+
+    return requestAsync(msg).then(res => {
+        if (res.value[0]) {
+
+            return JSON.parse(unicodeArray2Str(res.value[0]));
+        }
+    });
+   
+};
+
+/**
+ * 获取邀请码
+ */
+export const getInviteCode = () => {
+    const msg = { type: 'wallet/cloud@get_invite_code', param: {} };
+
+    return requestAsync(msg);
 };
