+++ src/app/public/interface.ts
